% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_calc.R
\name{my_calc}
\alias{my_calc}
\title{A calculator function}
\usage{
my_calc(x, y, fun)
}
\arguments{
\item{x}{A number, vector, or matrix}

\item{y}{A number, vector, or matrix}

\item{fun}{A string with an operation to be performed: sum, subtract,
divide, multiply}
}
\description{
This function is a calculator that takes vectors and matrices
# as input
}
\examples{
my_calc(7, 5, "sum")
my_calc(10:15, 1:5, "divide")
 my_calc(2, 5, "multiply")
}
