% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_multiplier.R
\name{my_multiplier}
\alias{my_multiplier}
\title{A multiplying function}
\usage{
my_multiplier(value)
}
\arguments{
\item{value}{A number}
}
\description{
This function multiples numbers between 5 and 10 by 5 and
numbers greater than 10 by 10
}
\examples{
my_multiplier(1)
my_multiplier(7)
my_multiplier(11)
}
